[hueemulation] Remove base type restriction from group items (#4428)
* Remove base type restriction from group items

Given that this is not documented anywhere and given that the user already
understands what they're doing when adding one of the hueemulation-used
tags, restricting them to add a base type to a group doesn't seem to be
reasonable. As this restriction makes debugging more complicated ("Why
does my group with the Lighting tag does not get discovered by
hueemulation?") this commit removes this restriction completely. As a
result, all groups with one of the supported tags, will be added to
hueemulation discovery document.

Fixes #4379

Signed-off-by: Florian <florian.schmidt.welzow@t-online.de>
Signed-off-by: Pascal Larin <plarin@gmail.com>